+++ base/admin.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+from .models import User, Student, Staff
+# Register your models here.
+
+admin.site.register(User)
+admin.site.register(Student)
+admin.site.register(Staff)
