--- students/admin.py
@@ -1,6 +1,0 @@
-from django.contrib import admin
-from .models import Student
-
-# Register your models here.
-
-admin.site.register(Student)
