--- users/admin.py
@@ -1,7 +1,0 @@
-from django.contrib import admin
-from django.contrib.auth.admin import UserAdmin
-from .models import User
-
-# Register your models here.
-
-admin.site.register(User, UserAdmin)
